Don't fail if test report does not exist

+++ test/spec/upload_test_reports.rb
@@ -19,12 +19,22 @@ time = Time.now.strftime '%Y%m%d%H%M%S'
 
 if ARGV[0] == 'IVT'
   logger.info("Uploading ivttest-#{time}.xml to COS")
+  unless File.exists?("#{Dir.pwd}/ivttest.xml")
+    logger.info('ivttest.xml not found. Skipping test report upload.')
+    exit 0
+  end
+
   doc = Nokogiri::XML(File.open("#{Dir.pwd}/ivttest.xml")) { |file| file.noblanks }
   doc.search('//testsuite').attribute('name').value = "hri-flink-validation-passthrough - #{ENV['BRANCH_NAME']} - IVT"
   File.rename("#{Dir.pwd}/ivttest.xml", "#{Dir.pwd}/hri-flink-validation-passthrough-ivttest-#{time}.xml")
   cos_helper.upload_object_data('wh-hri-dev1-allure-reports', "hri-flink-validation-passthrough-ivttest-#{time}.xml", File.read(File.join(Dir.pwd, "hri-flink-validation-passthrough-ivttest-#{time}.xml")))
 elsif ARGV[0] == 'Nightly'
   logger.info("Uploading nightlytest-#{time}.xml to COS")
+  unless File.exists?("#{Dir.pwd}/nightlytest.xml")
+    logger.info('nightlytest.xml not found. Skipping test report upload.')
+    exit 0
+  end
+
   doc = Nokogiri::XML(File.open("#{Dir.pwd}/nightlytest.xml")) { |file| file.noblanks }
   doc.search('//testsuite').attribute('name').value = "hri-flink-validation-passthrough - #{ENV['BRANCH_NAME']} - Nightly"
   File.rename("#{Dir.pwd}/nightlytest.xml", "#{Dir.pwd}/hri-flink-validation-passthrough-nightlytest-#{time}.xml")
